* rewritten network buffer code. now it uses a "bip buffer"-concept based circular buffer. larger servers will see a large benefit from this, as it effectively removes all memmove()'s from the network code, apart from the periodic memmove() to optimize the buffer for maximum space. - removed select/poll socket engines, i never finished them and if you're running FBSD 4.x / linux 2.4, then you need to upgrade anyway - unix compile still needs to be fixed
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4346 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/SpellAuras.cpp
@@ -2628,7 +2628,7 @@ void Aura::SpellAuraModStealth(bool apply)
 			cd.spellid = m_spellProto->Id;
 			static_cast<Player*>(m_target)->GetSession()->OutPacket( SMSG_COOLDOWN_EVENT, sizeof(packetSMSG_COOLDOWN_EVENT), &cd);
 			if( ((Player*)m_target)->m_outStealthDamageBonusPeriod && ((Player*)m_target)->m_outStealthDamageBonusPct )
-				((Player*)m_target)->m_outStealthDamageBonusTimer = UNIXTIME + ((Player*)m_target)->m_outStealthDamageBonusPeriod;
+				((Player*)m_target)->m_outStealthDamageBonusTimer = (uint32)UNIXTIME + ((Player*)m_target)->m_outStealthDamageBonusPeriod;
 		}
 	}
